Refactor to support dynamic authentication provider endpoints
