+++ config/routes.rb
@@ -66,7 +66,7 @@ Rails.application.routes.draw do
       get     "/resources/:account"                   => "resources#index"
       get     "/resources"                            => "resources#index"
 
-      get     "/authn-oidc/:account/providers"              => "authenticators/oidc/providers#index"
+      get     "/:authenticator/:account/providers"  => "providers#index"
 
       # NOTE: the order of these routes matters: we need the expire
       #       route to come first.
